+++ packages/cli/src/docs/handler.ts
@@ -4,12 +4,15 @@ import docgen from '../docgen/handler';
 import { Opts } from '../commands';
 import { createNullLogger } from '../util/logger';
 import type { Logger } from '../util/logger';
-import type { FunctionDescription } from '@openfn/describe-package';
+import type {
+  FunctionDescription,
+  PackageDescription,
+} from '@openfn/describe-package';
 
 import { getNameAndVersion, getLatestVersion } from '@openfn/runtime';
 import expandAdaptors from '../util/expand-adaptors';
 
-const describe = (adaptorName: string, fn: FunctionDescription) => `## ${
+const describeFn = (adaptorName: string, fn: FunctionDescription) => `## ${
   fn.name
 }(${fn.parameters.map(({ name }) => name).join(',')})
 
@@ -17,7 +20,18 @@ ${fn.description}
 
 ### Usage Examples
 
-${fn.examples.length ? fn.examples.map((eg) => eg).join('\n\n') : 'None'}
+${
+  fn.examples.length
+    ? fn.examples
+        .map(({ code, caption }) => {
+          if (caption) {
+            return `${caption}:\n${code}`;
+          }
+          return code;
+        })
+        .join('\n\n')
+    : 'None'
+}
 
 ### API Reference
 
@@ -25,6 +39,17 @@ https://docs.openfn.org/adaptors/packages/${adaptorName.replace(
   '@openfn/language-',
   ''
 )}-docs#${fn.name}
+`;
+
+const describeLib = (
+  adaptorName: string,
+  data: PackageDescription
+) => `## ${adaptorName} ${data.version}
+
+${data.functions
+  .map((fn) => `  ${fn.name}(${fn.parameters.map((p) => p.name).join(', ')})`)
+  .sort()
+  .join('\n')}
 `;
 
 const docsHandler = async (
@@ -60,12 +85,18 @@ const docsHandler = async (
     const source = await readFile(path, 'utf8');
     const data = JSON.parse(source);
 
-    const fn = data.functions.find(({ name }) => name === operation);
-    logger.debug('Operation schema:', fn);
-    logger.success(`Documentation for ${name}.${operation} v${version}:\n`);
+    let desc;
+    if (operation) {
+      const fn = data.functions.find(({ name }) => name === operation);
+      logger.debug('Operation schema:', fn);
+      logger.success(`Documentation for ${name}.${operation} v${version}:\n`);
 
-    // Generate a documentation string
-    const desc = describe(name, fn);
+      // Generate a documentation string
+      desc = describeFn(name, fn);
+    } else {
+      logger.debug('No operation provided, listing available operations');
+      desc = describeLib(name, data);
+    }
     // Log the description without any ceremony/meta stuff from the logger
     logger.print(desc);